Move resources to testclasses in filters
